Создаю Generic Stack
