+++ src/main/java/org/battle/models/AbstractWeapon.java
@@ -1,46 +1,21 @@
 package org.battle.models;
 
-import lombok.AllArgsConstructor;
-import org.battle.exceptions.StackFullException;
-
-@AllArgsConstructor
-
 public abstract class AbstractWeapon {
     // Переменная, содержащая максимальное количество патронов в магазине
-    private final int maxClipSize;
+    private int maxClipSize;
     // Переменная, содержащая тип стрельбы оружия
-    private final ShootType shootType;
+    private ShootType shootType;
     // Переменная, содержащая обойму
     private Stack<Ammo> clip;
     // Переменная, содержащая признак, что обойма пуста
     private boolean isEmptyClip = clip.isEmpty();
 
     // Метод создания патрона
-    private Ammo createAmmo(String ammoType) {
-        switch (ammoType) {
-            case "SELF_MADE":
-                return Ammo.SELF_MADE;
-            case "ARMOR_PIERSING":
-                return Ammo.ARMOR_PIERSING;
-            default:
-                return Ammo.DEFAULT;
-        }
-    }
+    public abstract Ammo createAmmo(String ammoType);
 
     // Метод перезарядки
-    public void reload(String ammoType) {
-        if (isEmptyClip) {
-            for (int i = 0; i < maxClipSize; i++) {
-                try {
-                    clip.push(createAmmo(ammoType));
-                } catch (StackFullException e) {
-                    // TODO catch body
-                }
-            }
-        } //TODO else
-    }
+    public abstract void reload(String ammoType);
+
     // Метод получения патрона для выстрела
-    public Ammo getAmmoForShoot() {
-        return clip.getAmmo();
-    }
+    public abstract Ammo getAmmoForShoot();
 }
