+++ src/main/java/org/battle/models/Weapons.java
@@ -3,16 +3,8 @@ package org.battle.models;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-@Getter
-public final class Weapons {
-    private static Weapons instance;
-
-    private Weapons() {}
-
-    public static Weapons getInstance() {
-        if (instance == null) {
-            instance = new Weapons();
-        }
-        return instance;
-    }
+// Фабрика
+public abstract class Weapons {
+    // Фабричный метод
+    public abstract AbstractWeapon createWeapon();
 }
